Created a new class to handle the Secondary View and removed the code from Mainpage
