+++ Fastedit/Views/MainPage.xaml.cs
@@ -47,6 +47,7 @@ namespace Fastedit
         private readonly SaveFileHelper savefilehelper = new SaveFileHelper();
         private readonly TabPageHelper tabpagehelper = new TabPageHelper();
         private readonly CustomDesigns customdesigns = null;
+        public readonly SecondaryView secondaryviews = null;
 
         //Settings entrys:
         private FontFamily TextBoxFontfamily = new FontFamily(DefaultValues.DefaultFontFamily);
@@ -92,7 +93,6 @@ namespace Fastedit
         private TextControlBox CurrentlySelectedTabPage_Textbox = null;
         private muxc.TabViewItem CurrentlySelectedTabPage = null;
         private muxc.FontIconSource TabPageFontIconSource = null;
-        private List<OpenedSecondaryViewItem> OpenedSecondaryViews = new List<OpenedSecondaryViewItem>();
         private muxc.TabViewItem SettingsTabPage = null;
         private NavigationEventArgs navigaionEvent = null;
 
@@ -104,7 +104,8 @@ namespace Fastedit
                 tabactions = new TabActions(TextTabControl, this);
             if(customdesigns == null)
                 customdesigns = new CustomDesigns(null, this);
-
+            if (secondaryviews == null)
+                secondaryviews = new SecondaryView(this, tabactions, TextTabControl);
             //Subscribe to the events:
             SizeChanged += MainPage_SizeChanged;
             SystemNavigationManagerPreview.GetForCurrentView().CloseRequested += Application_OnCloseRequest;
@@ -405,7 +406,7 @@ namespace Fastedit
             if (IsContentDialogOpen())
                 e.Handled = true;
 
-            if (OpenedSecondaryViews.Count > 0)
+            if (secondaryviews.OpenedSecondaryViews.Count > 0)
             {
                 ShowInfobar(InfoBarMessages.CloseAllInstances, InfoBarMessages.CloseAllInstances_Title, muxc.InfoBarSeverity.Warning);
                 e.Handled = true;
@@ -603,7 +604,7 @@ namespace Fastedit
         private async void TextTabControl_TabDroppedOutside(muxc.TabView sender, muxc.TabViewTabDroppedOutsideEventArgs args)
         {
             if (args.Tab != null)
-                await ExpandTabPageToNewView(args.Tab);
+                await secondaryviews.ExpandTabPageToNewView(args.Tab);
         }
 
         //Titlebar
@@ -1432,7 +1433,7 @@ namespace Fastedit
         private async void ExpandTabToNewWindow_Action()
         {
             if (CurrentlySelectedTabPage != null)
-                await ExpandTabPageToNewView(CurrentlySelectedTabPage);
+                await secondaryviews.ExpandTabPageToNewView(CurrentlySelectedTabPage);
         }
         private void Fullscreen_Action()
         {
@@ -2143,160 +2144,6 @@ namespace Fastedit
             }
         }
 
-        //Functions to handle the secondary windows
-        private async Task<int> GetViewIndexFromAppView(int ViewId)
-        {
-            int returnval = -1;
-            for (int i = 0; i < OpenedSecondaryViews.Count; i++)
-            {
-                await OpenedSecondaryViews[i].CoreApplicationView.Dispatcher.RunAsync(CoreDispatcherPriority.Normal, () =>
-                {
-                    if (ApplicationView.GetForCurrentView().Id == ViewId)
-                    {
-                        returnval = i;
-                    }
-                });
-            }
-            return returnval;
-        }
-        public async Task ExpandTabPageToNewView(muxc.TabViewItem TabPage)
-        {
-            var textcontrolbox = tabactions.GetTextBoxFromTabPage(TabPage);
-
-            //Run code only if tabpage is a textbox:
-            if (textcontrolbox == null) return;
-
-            string TabPageName = TabPage.Name;
-            string Header = textcontrolbox.Header;
-            string tbtext = textcontrolbox.GetText();
-            double tbZoomFactor = textcontrolbox._zoomFactor;
-            TextWrapping wrapping = textcontrolbox.WordWrap;
-            bool IsReadOnly = textcontrolbox.IsReadOnly;
-            int SelectionStart = textcontrolbox.SelectionStart;
-            ApplicationView appview = null;
-            CoreApplicationView newView = CoreApplication.CreateNewView();
-
-            int newViewId = 0;
-            await newView.Dispatcher.RunAsync(CoreDispatcherPriority.Normal, () =>
-            {
-                Frame frame = new Frame();
-                frame.Navigate(typeof(TextboxViewPage));
-                Window.Current.Content = frame;
-                Window.Current.Activate();
-
-                if (frame.Content is TextboxViewPage tbviewpage)
-                {
-                    tbviewpage.TabPageName = TabPageName;
-                    tbviewpage.Title = Header;
-                    tbviewpage.TextboxText = tbtext;
-                    tbviewpage.SelectionStart = SelectionStart;
-                    tbviewpage.Textwrapping = wrapping;
-                    tbviewpage.ZoomFactor = tbZoomFactor;
-                    tbviewpage.IsReadonly = IsReadOnly;
-                }
-
-                appview = ApplicationView.GetForCurrentView();
-                newViewId = appview.Id;
-                appview.Consolidated += TextboxViewPage_Consolidated;
-            });
-
-            TabPage.Visibility = Visibility.Collapsed;
-
-            //if all tabs are in a secondary view, create a new tab
-            if (tabactions.GetShownTabPages() < 1)
-            {
-                muxc.TabViewItem tab = tabactions.NewTab();
-                if (tab != null)
-                {
-                    TextTabControl.SelectedItem = tab;
-                    await tabactions.SaveAllTabChanges();
-                    SetSettingsToTabPage(tab, TextBoxMargin());
-                }
-            }
-
-            await tabactions.GetTextBoxFromTabPage(TabPage).SetText("Cleared the textbox");
-            for (int i = 0; i < TextTabControl.TabItems.Count; i++)
-            {
-                if (TextTabControl.TabItems[i] is muxc.TabViewItem Tab)
-                {
-                    if (Tab.Visibility == Visibility.Visible)
-                    {
-                        TextTabControl.SelectedItem = Tab;
-                        break;
-                    }
-                }
-            }
-
-            if (!await ApplicationViewSwitcher.TryShowAsStandaloneAsync(newViewId))
-            {
-                ShowInfobar(InfoBarMessages.CouldNotOpenInNewView, InfoBarMessages.CouldNotOpenInNewViewTitle,  muxc.InfoBarSeverity.Error);
-                return;
-            }
-            OpenedSecondaryViews.Add(new OpenedSecondaryViewItem
-            {
-                ApplicationView = appview,
-                CoreApplicationView = newView
-            });
-        }
-        private async Task<bool> CloseExpandedView(ApplicationView sender, bool RemoveFromOpenedViewList = true)
-        {
-            int index = await GetViewIndexFromAppView(sender.Id);
-            if (index < 0 || index >= OpenedSecondaryViews.Count)
-                return false;
-
-            TextboxViewPage tbvpage = null;
-            string TabPageName = "";
-
-            //Get the view by the index returned from the GetViewIndexFromAppView
-            var view = OpenedSecondaryViews[index];
-            await view.CoreApplicationView.Dispatcher.RunAsync(CoreDispatcherPriority.Normal, () =>
-            {
-                if (Window.Current.Content is Frame frame)
-                {
-                    if (frame.Content is TextboxViewPage tbvp)
-                    {
-                        tbvpage = tbvp;
-                        TabPageName = tbvpage.TabPageName;
-                    }
-                }
-            });
-            if (tbvpage == null)
-                return false;
-            double ZoomFactor = tbvpage.ZoomFactor;
-            bool IsReadOnly = tbvpage.IsReadonly;
-            string TextBoxText = tbvpage.TextboxText;
-            int SelectionStart = tbvpage.SelectionStart;
-            bool IsModified = tbvpage.IsModified;
-
-            if (tbvpage != null)
-            {
-                await this.Dispatcher.RunAsync(CoreDispatcherPriority.Normal, async () =>
-                {
-                    if (TextTabControl.FindName(TabPageName) is muxc.TabViewItem tabpage)
-                    {
-                        var textbox = tabactions.GetTextBoxFromTabPage(tabpage);
-                        if (textbox != null)
-                        {
-                            await textbox.SetText(TextBoxText);
-                            textbox.SetFontZoomFactor(ZoomFactor);
-                            textbox.IsReadOnly = IsReadOnly;
-                            textbox.SelectionStart = SelectionStart;
-                            tabpagehelper.SetTabModified(tabpage, IsModified);
-                        }
-                        tabpage.Visibility = Visibility.Visible;
-                        TextTabControl.SelectedItem = tabpage;
-                    }
-                });
-                if (RemoveFromOpenedViewList)
-                    return OpenedSecondaryViews.Remove(view);
-                return true;
-            }
-            return false;
-        }
-        private async void TextboxViewPage_Consolidated(ApplicationView sender, ApplicationViewConsolidatedEventArgs args)
-        {
-            await CloseExpandedView(sender);
-        }
     }
     
     public class SettingsNavigationParameter
